[Fix] (ZU) float to int for pourcentageConstructible arg in constructor

+++ include/ZU.hpp
@@ -8,7 +8,7 @@ class ZU: public Constructible{
     private:
         float surfaceConstruite;
     public:
-    ZU(int num, std::string prop, Polygone<float> forme, float surfaceConstruite, float pourcentageConstructible);
+    ZU(int num, std::string prop, Polygone<float> forme, float surfaceConstruite, int pourcentageConstructible);
     ZU(const ZU & zu);
     ~ZU();
